Split high into separate libraries
