+++ Doxyfile
@@ -179,15 +179,18 @@ STRIP_FROM_PATH        =
 # specify the list of include paths that are normally passed to the compiler
 # using the -I flag.
 
-STRIP_FROM_INC_PATH    = type_traits/include \
-                         high/include \
+STRIP_FROM_INC_PATH    = functional/include \
                          literal/include \
                          log/include \
                          math/include \
                          numeric/include \
+                         optional/include \
                          random/include \
                          range/include \
-                         type_support/include
+                         tuple/include \
+                         type_support/include \
+                         type_traits/include \
+                         utility/include
 
 # If the SHORT_NAMES tag is set to YES, doxygen will generate much shorter (but
 # less readable) file names. This can be useful is your file systems doesn't
@@ -852,8 +855,7 @@ WARN_LOGFILE           =
 # spaces. See also FILE_PATTERNS and EXTENSION_MAPPING
 # Note: If this tag is empty the current directory is searched.
 
-INPUT                  = high/include/vccc \
-                         literal/include/vccc \
+INPUT                  = literal/include/vccc \
                          log/include/vccc \
                          math/include/vccc \
                          numeric/include/vccc \
