+++ functional/include/vccc/functional.hpp
@@ -1,0 +1,15 @@
+# /*
+#  * Created by YongGyu Lee on 2020/12/08.
+#  */
+#
+# ifndef VCCC_FUNCTIONAL_FUNCTIONAL_HPP
+# define VCCC_FUNCTIONAL_FUNCTIONAL_HPP
+#
+# include "vccc/functional/invoke.hpp"
+
+/**
+@defgroup functional functional
+*/
+
+
+# endif // VCCC_FUNCTIONAL_FUNCTIONAL_HPP
