+++ functional/test/CMakeLists.txt
@@ -1,0 +1,1 @@
+VCCC_TEST(functional_test functional_test.cpp VCCC::functional)
