+++ functional/test/test_core.hpp
