--- high/include/vccc/high.hpp
@@ -1,15 +1,0 @@
-# /*
-#  * Created by YongGyu Lee on 2020/12/08.
-#  */
-#
-# ifndef VCCC_HIGH_HPP
-# define VCCC_HIGH_HPP
-#
-# include "vccc/high/functional.hpp"
-# include "vccc/high/tuple.hpp"
-
-/**
-@defgroup high high
-*/
-
-# endif //VCCC_HIGH_HPP
