--- high/include/vccc/high/functional.hpp
@@ -1,18 +1,0 @@
-# /*
-#  * Created by YongGyu Lee on 2020/12/08.
-#  */
-#
-# ifndef VCCC_HIGH_FUNCTIONAL_HPP
-# define VCCC_HIGH_FUNCTIONAL_HPP
-#
-# include "vccc/high/functional/invoke.hpp"
-
-/**
-@addtogroup high
-@{
-@defgroup high_functional functional
-@}
-*/
-
-
-# endif //VCCC_HIGH_FUNCTIONAL_HPP
