--- high/include/vccc/high/tuple.hpp
@@ -1,17 +1,0 @@
-# /*
-#  * Created by YongGyu Lee on 2020/12/08.
-#  */
-#
-# ifndef VCCC_HIGH_TUPLE_HPP
-# define VCCC_HIGH_TUPLE_HPP
-#
-# include "vccc/high/tuple/apply.hpp"
-
-/**
-@addtogroup high
-@{
-@defgroup high_tuple tuple
-@}
-*/
-
-# endif //VCCC_HIGH_TUPLE_HPP
