--- high/test/CMakeLists.txt
@@ -1,1 +1,0 @@
-VCCC_TEST(high_functional functional_test.cpp VCCC::high)
