+++ tuple/include/vccc/tuple.hpp
@@ -1,0 +1,14 @@
+# /*
+#  * Created by YongGyu Lee on 2020/12/08.
+#  */
+#
+# ifndef VCCC_TUPLE_HPP
+# define VCCC_TUPLE_HPP
+#
+# include "vccc/tuple/apply.hpp"
+
+/**
+@defgroup tuple tuple
+*/
+
+# endif //VCCC_TUPLE_HPP
